requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiohttp-cors==0.8.1
aiohttp-jinja2==1.6
aiohttp-session==2.12.1
aiokafka==0.12.0
aiosignal==1.3.2
altgraph==0.17.4
ansi2html==1.9.2
ansi2image==0.1.4
appdirs==1.4.4
asttokens==3.0.0
async-timeout==5.0.1
attrs==24.3.0
Automat==24.8.1
bcrypt==5.0.0
beautifulsoup4==4.12.3
beautifultable==1.1.0
blinker==1.9.0
cachetools==5.5.2
cairocffi==1.7.1
CairoSVG==2.8.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
configparser==7.2.0
constantly==23.10.4
contourpy==1.3.1
coverage==6.5.0
coveralls==3.3.1
cryptography==43.0.3
cssselect2==0.8.0
cycler==0.12.1
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
et_xmlfile==2.0.0
executing==2.2.0
Flask==3.1.0
fonttools==4.56.0
frozendict==2.4.6
frozenlist==1.5.0
future==1.0.0
genson==1.3.0
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.2
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
h11==0.14.0
httplib2==0.22.0
hyperlink==21.0.0
hypothesis==6.129.4
idna==3.10
imgkit==1.2.3
importlib_metadata==8.6.1
incremental==24.7.2
iniconfig==2.1.0
ipython==9.0.2
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lxml==5.3.1
macholib==1.16.3
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mock==5.2.0
MouseInfo==0.1.3
multidict==6.2.0
nautilus==0.5.2
nltk==3.9.1
nose2==0.15.1
nulltype==2.3.1
numpy==2.2.4
oauthlib==3.2.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
peewee==3.18.2
pexpect==4.9.0
pillow==11.1.0
plaid-python==37.0.0
pluggy==1.5.0
prompt_toolkit==3.0.50
propcache==0.3.0
proto-plus==1.26.1
protobuf==6.30.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pycparser==2.22
pyee==11.1.1
pygal==3.0.5
PyGetWindow==0.0.9
Pygments==2.19.1
pyinstaller-hooks-contrib==2025.1
PyJWT==2.10.1
PyMsgBox==2.0.1
pyobjc-core==11.1
pyobjc-framework-Cocoa==11.1
pyobjc-framework-Quartz==11.1
pyOpenSSL==25.0.0
pyotp==2.9.0
pyparsing==3.2.1
pyperclip==1.11.0
pyppeteer==2.0.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pyte==0.8.2
pytest==8.3.5
pytest-cov==5.0.0
python-consul==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytweening==1.2.0
pytz==2025.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
robin-stocks==3.4.0
rpds-py==0.23.1
rsa==4.9
rubicon-objc==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.28.1
setuptools==76.1.0
simplejson==3.20.1
six==1.17.0
snaptrade-python-sdk==11.0.82
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
tabpy==2.13.0
tabulate==0.9.0
termtosvg==1.1.0
textblob==0.19.0
threadpoolctl==3.6.0
tinycss2==1.4.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.28.0
trio-websocket==0.11.1
Twisted==24.11.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==1.26.20
uvloop==0.21.0
wcwidth==0.2.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
zope.interface==7.2
